income-platform/src/agent-14-data-quality/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
psycopg2-binary==2.9.9
pydantic-settings==2.5.2
httpx==0.27.2
python-jose[cryptography]==3.3.0
